+++ go-app/config/config_interface.go
@@ -1,0 +1,7 @@
+package config
+
+type DeviceConfiguration interface {
+	GetUsbID() string
+	GetUniqueID() string
+	Matches(c DeviceConfiguration) bool
+}