+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 
 # CHANGELOG
 
+## v1.15.0
+- Add option to calibrate using the unique device ID instead of the standard device ID (useful for arduino's and other similar devices)
+- Improve direct control server restart handling and error logging
+- Website security updates
+
 ## v1.14.3
 - Add notice about Cab Debugger functionality